Renaming to make commandline callable
Added to cmdline path, but not cmdline inputs currently
