+++ nih2mne/utilities/make_meg_bids_fromcsv.py
