+++ setup.py
@@ -22,5 +22,6 @@ setuptools.setup(
     install_requires=['mne', 'numpy', 'pytest', 'joblib', 'nibabel','mne_bids','pandas'],
     scripts=['nih2mne/bstags.py',
         'nih2mne/calc_mnetrans.py',
-        'nih2mne/make_meg_bids.py'],
+        'nih2mne/make_meg_bids.py',
+        'nih2mn3/utilities/make_meg_bids_fromcsv.py],
 )
